🖋 chore: 멀티 모듈 형태로 분리
