+++ .gitignore
@@ -42,5 +42,7 @@ out/
 ### properties ###
 *.properties
 *.yml
-/src/main/resources/Wallet_kuberLoudy/
+/kl-api/src/main/resources/Wallet_kuberLoudy/
 
+
+kl-api/src/main/resources/application.yaml
