+++ kl-api/src/main/java/com/kuberloudy/api/global/common/ResponseAdvice.java
@@ -1,0 +1,36 @@
+package com.kuberloudy.api.global.common;
+
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.core.MethodParameter;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.server.ServerHttpRequest;
+import org.springframework.http.server.ServerHttpResponse;
+import org.springframework.http.server.ServletServerHttpResponse;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseBodyAdvice;
+
+@RestControllerAdvice
+public class ResponseAdvice implements ResponseBodyAdvice {
+    @Override
+    public boolean supports(MethodParameter returnType, Class converterType) {
+        return false;
+    }
+
+    @Override
+    public Object beforeBodyWrite(Object body, MethodParameter returnType,
+                                  MediaType selectedContentType, Class selectedConverterType,
+                                  ServerHttpRequest request, ServerHttpResponse response)
+    {
+        HttpServletResponse servletResponse = ((ServletServerHttpResponse) response).getServletResponse();
+
+        int code = servletResponse.getStatus();
+        HttpStatus httpStatus = HttpStatus.resolve(code);
+
+        if (httpStatus == null) {
+            return body;
+        }
+
+        return null;
+    }
+}
