+++ kl-api/src/main/java/com/kuberloudy/api/iam/controller/dto/IamUser.java
@@ -1,0 +1,9 @@
+package com.kuberloudy.api.iam.controller.dto;
+
+
+import lombok.Getter;
+
+@Getter
+public class IamUser {
+    private String iamId;
+}
