+++ kl-api/src/main/java/com/kuberloudy/api/iam/controller/dto/IamUserKey.java
@@ -1,0 +1,9 @@
+package com.kuberloudy.api.iam.controller.dto;
+
+import lombok.Getter;
+
+@Getter
+public class IamUserKey {
+    private String accessKey;
+    private String secretKey;
+}
