+++ kl-api/src/main/java/com/kuberloudy/api/iam/repository/IamRepository.java
@@ -1,0 +1,7 @@
+package com.kuberloudy.api.iam.repository;
+
+import com.kuberloudy.api.iam.entity.Iam;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface IamRepository extends JpaRepository<Iam,Long> {
+}
