+++ kl-api/src/main/java/com/kuberloudy/api/member/entity/Password.java
@@ -1,0 +1,29 @@
+package com.kuberloudy.api.member.entity;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(name = "member_password")
+public class Password {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long passwordId;
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
+    public Member member;
+
+    @Column(length = 255)
+    public String password;
+
+    @Builder
+    public Password(Member member, String password) {
+        this.member = member;
+        this.password = password;
+    }
+}
