+++ kl-api/src/main/java/com/kuberloudy/api/member/entity/Provider.java
@@ -1,0 +1,7 @@
+package com.kuberloudy.api.member.entity;
+
+public enum Provider {
+    GOOGLE,
+    GITHUB,
+    KUBERLOUDY
+}
