+++ kl-api/src/main/java/com/kuberloudy/api/member/repository/MemberRepository.java
@@ -1,0 +1,7 @@
+package com.kuberloudy.api.member.repository;
+
+import com.kuberloudy.api.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member,Long> {
+}
