+++ kl-api/src/main/java/com/kuberloudy/api/member/service/MemberService.java
@@ -1,0 +1,4 @@
+package com.kuberloudy.api.member.service;
+
+public class MemberService {
+}
