+++ settings.gradle
@@ -1,1 +1,4 @@
 rootProject.name = 'spring'
+include 'kl-api'
+include 'kl-common'
+include 'kl-domain'
