--- src/main/java/com/kuberloudy/spring/domain/iam/controller/dto/IamUser.java
@@ -1,9 +1,0 @@
-package com.kuberloudy.spring.domain.iam.controller.dto;
-
-
-import lombok.Getter;
-
-@Getter
-public class IamUser {
-    private String iamId;
-}
