--- src/main/java/com/kuberloudy/spring/domain/iam/controller/dto/IamUserKey.java
@@ -1,9 +1,0 @@
-package com.kuberloudy.spring.domain.iam.controller.dto;
-
-import lombok.Getter;
-
-@Getter
-public class IamUserKey {
-    private String accessKey;
-    private String secretKey;
-}
