--- src/main/java/com/kuberloudy/spring/domain/member/controller/MemberController.java
@@ -1,12 +1,0 @@
-package com.kuberloudy.spring.domain.member.controller;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@Slf4j
-@RestController
-@RequestMapping("/members")
-public class MemberController {
-
-}
