--- src/main/java/com/kuberloudy/spring/domain/member/entity/Password.java
@@ -1,19 +1,0 @@
-package com.kuberloudy.spring.domain.member.entity;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "member_password")
-public class Password {
-
-    @Id
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    public Member member;
-
-    @Column(length = 255)
-    public String password;
-}
