--- src/main/java/com/kuberloudy/spring/domain/member/entity/Provider.java
@@ -1,7 +1,0 @@
-package com.kuberloudy.spring.domain.member.entity;
-
-public enum Provider {
-    GOOGLE,
-    GITHUB,
-    KUBERLOUDY
-}
