--- src/main/java/com/kuberloudy/spring/domain/member/repository/MemberRepository.java
@@ -1,7 +1,0 @@
-package com.kuberloudy.spring.domain.member.repository;
-
-import com.kuberloudy.spring.domain.member.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member,Long> {
-}
