--- src/test/java/com/kuberloudy/spring/ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.kuberloudy.spring;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
